Bug #1623905 unisoc: wlan: fix cfi issue in add_key
[root cause  ] fix cfi issue in add_key
[changes     ] drivers/unisoc_platform/wlan
[side effects] none
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:none
[confirm dependent commit]:none
[board]:sharkl3/sharkl5pro
[test case]:none
[change_type ] bringup --> sprd_chip_support
[tag_product ] common
[reviewers   ] keguang.zhang, baolei.yuan

This patch fixs cfi issue in add_key.

+++ drivers/unisoc_platform/wlan/common/cfg80211.c
@@ -227,7 +227,7 @@ static int cfg80211_add_cipher_key(struct sprd_vif *vif, bool pairwise,
 		ret = sprd_add_key(vif->priv, vif,
 				   vif->key[pairwise][key_index],
 				   vif->key_len[pairwise][key_index],
-				   pairwise, key_index, key_seq,
+				   (u8)pairwise, key_index, key_seq,
 				   *cipher_ptr, mac_addr);
 	}
